break up and reorganize patches to allow easier swithing between device types and pixel storage modes.
