+++ forcepho/patches/__init__.py
@@ -1,7 +1,16 @@
-from .patch import Patch
-from .jades_patch import JadesPatch
-from .static_patch import StaticPatch
+
+from .device_patch import GPUPatchMixin
+from .pixel_patch import StorePatch, FITSPatch
+
+
+class JadesPatch(StorePatch, GPUPatchMixin):
+    pass
+
+
+class SimplePatch(FITSPatch, GPUPatchMixin):
+    pass
+
 
 __all__ = ["Patch",
            "JadesPatch",
-           "StaticPatch"]
+           "SimplePatch"]
